initial commit -
TF module will create an azure sql server and an azure sql database in the same server.
The sql server will allow access to azure services.
The password for the sql server will be auto generated and returned as an input.
The resource group where all azure resources will be created is the Colony sandbox resource group.